chore: Shared text system on Linux and Windows
