+++ crates/gpui/src/platform/common.rs
@@ -1,0 +1,3 @@
+mod text_system;
+
+pub(crate) use text_system::*;
