+++ crates/gpui/src/platform/common/text_system.rs
@@ -5,11 +5,12 @@ use crate::{
 };
 use anyhow::{anyhow, Context, Ok, Result};
 use collections::HashMap;
-use cosmic_text::Font as CosmicTextFont;
 use cosmic_text::{
-    fontdb::Query, Attrs, AttrsList, BufferLine, CacheKey, Family, FontSystem, SwashCache,
+    Attrs, AttrsList, BufferLine, CacheKey, Family, Font as CosmicTextFont, FontSystem, SwashCache,
 };
-use parking_lot::{RwLock, RwLockUpgradableReadGuard};
+
+use itertools::Itertools;
+use parking_lot::RwLock;
 use pathfinder_geometry::{
     rect::{RectF, RectI},
     vector::{Vector2F, Vector2I},
@@ -17,45 +18,49 @@ use pathfinder_geometry::{
 use smallvec::SmallVec;
 use std::{borrow::Cow, sync::Arc};
 
-pub(crate) struct WindowsTextSystem(RwLock<WindowsTextSystemState>);
+pub(crate) struct CommonTextSystem(RwLock<CommonTextSystemState>);
 
-struct WindowsTextSystemState {
+struct CommonTextSystemState {
     swash_cache: SwashCache,
     font_system: FontSystem,
-    fonts: Vec<Arc<CosmicTextFont>>,
-    font_selections: HashMap<Font, FontId>,
-    font_ids_by_family_name: HashMap<SharedString, SmallVec<[FontId; 4]>>,
-    postscript_names_by_font_id: HashMap<FontId, String>,
+    /// Contains all already loaded fonts, including all faces. Indexed by `FontId`.
+    loaded_fonts_store: Vec<Arc<CosmicTextFont>>,
+    /// Caches the `FontId`s associated with a specific family to avoid iterating the font database
+    /// for every font face in a family.
+    font_ids_by_family_cache: HashMap<SharedString, SmallVec<[FontId; 4]>>,
+    /// The name of each font associated with the given font id
+    postscript_names: HashMap<FontId, String>,
 }
 
-impl WindowsTextSystem {
+impl CommonTextSystem {
     pub(crate) fn new() -> Self {
         let mut font_system = FontSystem::new();
-        Self(RwLock::new(WindowsTextSystemState {
+
+        // todo(linux) make font loading non-blocking
+        font_system.db_mut().load_system_fonts();
+
+        Self(RwLock::new(CommonTextSystemState {
             font_system,
             swash_cache: SwashCache::new(),
-            fonts: Vec::new(),
-            font_selections: HashMap::default(),
-            // font_ids_by_postscript_name: HashMap::default(),
-            font_ids_by_family_name: HashMap::default(),
-            postscript_names_by_font_id: HashMap::default(),
+            loaded_fonts_store: Vec::new(),
+            font_ids_by_family_cache: HashMap::default(),
+            postscript_names: HashMap::default(),
         }))
     }
 }
 
-impl Default for WindowsTextSystem {
+impl Default for CommonTextSystem {
     fn default() -> Self {
         Self::new()
     }
 }
 
-#[allow(unused)]
-impl PlatformTextSystem for WindowsTextSystem {
+impl PlatformTextSystem for CommonTextSystem {
     fn add_fonts(&self, fonts: Vec<Cow<'static, [u8]>>) -> Result<()> {
         self.0.write().add_fonts(fonts)
     }
 
-    // todo(windows) ensure that this integrates with platform font loading
+    // todo(linux) ensure that this integrates with platform font loading
     // do we need to do more than call load_system_fonts()?
     fn all_font_names(&self) -> Vec<String> {
         self.0
@@ -67,69 +72,63 @@ impl PlatformTextSystem for WindowsTextSystem {
             .collect()
     }
 
-    // todo(windows)
     fn all_font_families(&self) -> Vec<String> {
-        Vec::new()
+        self.0
+            .read()
+            .font_system
+            .db()
+            .faces()
+            // todo(linux) this will list the same font family multiple times
+            .filter_map(|face| face.families.first().map(|family| family.0.clone()))
+            .collect_vec()
     }
 
     fn font_id(&self, font: &Font) -> Result<FontId> {
-        // todo(windows): Do we need to use CosmicText's Font APIs? Can we consolidate this to use font_kit?
-        let lock = self.0.upgradable_read();
-        if let Some(font_id) = lock.font_selections.get(font) {
-            Ok(*font_id)
+        // todo(linux): Do we need to use CosmicText's Font APIs? Can we consolidate this to use font_kit?
+        let mut state = self.0.write();
+
+        let candidates = if let Some(font_ids) = state.font_ids_by_family_cache.get(&font.family) {
+            font_ids.as_slice()
         } else {
-            let mut lock = RwLockUpgradableReadGuard::upgrade(lock);
-            let candidates = if let Some(font_ids) = lock.font_ids_by_family_name.get(&font.family)
-            {
-                font_ids.as_slice()
-            } else {
-                let font_ids = lock.load_family(&font.family, font.features)?;
-                lock.font_ids_by_family_name
-                    .insert(font.family.clone(), font_ids);
-                lock.font_ids_by_family_name[&font.family].as_ref()
-            };
+            let font_ids = state.load_family(&font.family, font.features)?;
+            state
+                .font_ids_by_family_cache
+                .insert(font.family.clone(), font_ids);
+            state.font_ids_by_family_cache[&font.family].as_ref()
+        };
 
-            let id = lock
-                .font_system
-                .db()
-                .query(&Query {
-                    families: &[Family::Name(&font.family)],
-                    weight: font.weight.into(),
-                    style: font.style.into(),
-                    stretch: Default::default(),
-                })
-                .context("no font")?;
+        // todo(linux) ideally we would make fontdb's `find_best_match` pub instead of using font-kit here
+        let candidate_properties = candidates
+            .iter()
+            .map(|font_id| {
+                let database_id = state.loaded_fonts_store[font_id.0].id();
+                let face_info = state.font_system.db().face(database_id).expect("");
+                face_info_into_properties(face_info)
+            })
+            .collect::<SmallVec<[_; 4]>>();
 
-            let font_id = if let Some(font_id) = lock.fonts.iter().position(|font| font.id() == id)
-            {
-                FontId(font_id)
-            } else {
-                // Font isn't in fonts so add it there, this is because we query all the fonts in the db
-                // and maybe we haven't loaded it yet
-                let font_id = FontId(lock.fonts.len());
-                let font = lock.font_system.get_font(id).unwrap();
-                lock.fonts.push(font);
-                font_id
-            };
+        let ix =
+            font_kit::matching::find_best_match(&candidate_properties, &font_into_properties(font))
+                .context("requested font family contains no font matching the other parameters")?;
 
-            lock.font_selections.insert(font.clone(), font_id);
-            Ok(font_id)
-        }
+        Ok(candidates[ix])
     }
 
     fn font_metrics(&self, font_id: FontId) -> FontMetrics {
-        let metrics = self.0.read().fonts[font_id.0].as_swash().metrics(&[]);
+        let metrics = self.0.read().loaded_fonts_store[font_id.0]
+            .as_swash()
+            .metrics(&[]);
 
         FontMetrics {
             units_per_em: metrics.units_per_em as u32,
             ascent: metrics.ascent,
-            descent: -metrics.descent, // todo(windows) confirm this is correct
+            descent: -metrics.descent, // todo(linux) confirm this is correct
             line_gap: metrics.leading,
             underline_position: metrics.underline_offset,
             underline_thickness: metrics.stroke_size,
             cap_height: metrics.cap_height,
             x_height: metrics.x_height,
-            // todo(windows): Compute this correctly
+            // todo(linux): Compute this correctly
             bounding_box: Bounds {
                 origin: point(0.0, 0.0),
                 size: size(metrics.max_width, metrics.ascent + metrics.descent),
@@ -139,10 +138,11 @@ impl PlatformTextSystem for WindowsTextSystem {
 
     fn typographic_bounds(&self, font_id: FontId, glyph_id: GlyphId) -> Result<Bounds<f32>> {
         let lock = self.0.read();
-        let metrics = lock.fonts[font_id.0].as_swash().metrics(&[]);
-        let glyph_metrics = lock.fonts[font_id.0].as_swash().glyph_metrics(&[]);
+        let glyph_metrics = lock.loaded_fonts_store[font_id.0]
+            .as_swash()
+            .glyph_metrics(&[]);
         let glyph_id = glyph_id.0 as u16;
-        // todo(windows): Compute this correctly
+        // todo(linux): Compute this correctly
         // see https://github.com/servo/font-kit/blob/master/src/loaders/freetype.rs#L614-L620
         Ok(Bounds {
             origin: point(0.0, 0.0),
@@ -177,19 +177,19 @@ impl PlatformTextSystem for WindowsTextSystem {
         self.0.write().layout_line(text, font_size, runs)
     }
 
-    // todo(windows) Confirm that this has been superseded by the LineWrapper
+    // todo(linux) Confirm that this has been superseded by the LineWrapper
     fn wrap_line(
         &self,
-        text: &str,
-        font_id: FontId,
-        font_size: Pixels,
-        width: Pixels,
+        _text: &str,
+        _font_id: FontId,
+        _font_size: Pixels,
+        _width: Pixels,
     ) -> Vec<usize> {
         unimplemented!()
     }
 }
 
-impl WindowsTextSystemState {
+impl CommonTextSystemState {
     #[profiling::function]
     fn add_fonts(&mut self, fonts: Vec<Cow<'static, [u8]>>) -> Result<()> {
         let db = self.font_system.db_mut();
@@ -206,6 +206,7 @@ impl WindowsTextSystemState {
         Ok(())
     }
 
+    // todo(linux) handle `FontFeatures`
     #[profiling::function]
     fn load_family(
         &mut self,
@@ -226,27 +227,35 @@ impl WindowsTextSystemState {
                 .font_system
                 .get_font(font_id)
                 .ok_or_else(|| anyhow!("Could not load font"))?;
-            // TODO: figure out why this is causing fluent icons from loading
-            // if font.as_swash().charmap().map('m') == 0 {
-            //     self.font_system.db_mut().remove_face(font.id());
-            //     continue;
-            // };
 
-            let font_id = FontId(self.fonts.len());
+            // HACK: To let the storybook run and render Windows caption icons. We should actually do better font fallback.
+            let allowed_bad_font_names = vec![
+                "SegoeFluentIcons", // NOTE: Segoe fluent icons postscript name is inconsistent
+                "Segoe Fluent Icons",
+            ];
+
+            if font.as_swash().charmap().map('m') == 0
+                && !allowed_bad_font_names.contains(&postscript_name.as_str())
+            {
+                self.font_system.db_mut().remove_face(font.id());
+                continue;
+            };
+
+            let font_id = FontId(self.loaded_fonts_store.len());
             font_ids.push(font_id);
-            self.fonts.push(font);
-            self.postscript_names_by_font_id
-                .insert(font_id, postscript_name);
+            self.loaded_fonts_store.push(font);
+            self.postscript_names.insert(font_id, postscript_name);
         }
+
         Ok(font_ids)
     }
 
     fn advance(&self, font_id: FontId, glyph_id: GlyphId) -> Result<Size<f32>> {
-        let width = self.fonts[font_id.0]
+        let width = self.loaded_fonts_store[font_id.0]
             .as_swash()
             .glyph_metrics(&[])
             .advance_width(glyph_id.0 as u16);
-        let height = self.fonts[font_id.0]
+        let height = self.loaded_fonts_store[font_id.0]
             .as_swash()
             .glyph_metrics(&[])
             .advance_height(glyph_id.0 as u16);
@@ -254,7 +263,10 @@ impl WindowsTextSystemState {
     }
 
     fn glyph_for_char(&self, font_id: FontId, ch: char) -> Option<GlyphId> {
-        let glyph_id = self.fonts[font_id.0].as_swash().charmap().map(ch);
+        let glyph_id = self.loaded_fonts_store[font_id.0]
+            .as_swash()
+            .charmap()
+            .map(ch);
         if glyph_id == 0 {
             None
         } else {
@@ -263,17 +275,14 @@ impl WindowsTextSystemState {
     }
 
     fn is_emoji(&self, font_id: FontId) -> bool {
-        // todo(windows): implement this correctly
-        self.postscript_names_by_font_id
+        // TODO: Include other common emoji fonts
+        self.postscript_names
             .get(&font_id)
-            .map_or(false, |postscript_name| {
-                postscript_name == "AppleColorEmoji"
-            })
+            .map_or(false, |postscript_name| postscript_name == "NotoColorEmoji")
     }
 
-    // todo(windows) both raster functions have problems because I am not sure this is the correct mapping from cosmic text to gpui system
     fn raster_bounds(&mut self, params: &RenderGlyphParams) -> Result<Bounds<DevicePixels>> {
-        let font = &self.fonts[params.font_id.0];
+        let font = &self.loaded_fonts_store[params.font_id.0];
         let font_system = &mut self.font_system;
         let image = self
             .swash_cache
@@ -305,9 +314,9 @@ impl WindowsTextSystemState {
         if glyph_bounds.size.width.0 == 0 || glyph_bounds.size.height.0 == 0 {
             Err(anyhow!("glyph bounds are empty"))
         } else {
-            // todo(windows) handle subpixel variants
+            // todo(linux) handle subpixel variants
             let bitmap_size = glyph_bounds.size;
-            let font = &self.fonts[params.font_id.0];
+            let font = &self.loaded_fonts_store[params.font_id.0];
             let font_system = &mut self.font_system;
             let image = self
                 .swash_cache
@@ -329,17 +338,43 @@ impl WindowsTextSystemState {
         }
     }
 
-    // todo(windows) This is all a quick first pass, maybe we should be using cosmic_text::Buffer
+    fn font_id_for_cosmic_id(&mut self, id: cosmic_text::fontdb::ID) -> FontId {
+        if let Some(ix) = self
+            .loaded_fonts_store
+            .iter()
+            .position(|font| font.id() == id)
+        {
+            FontId(ix)
+        } else {
+            // This matches the behavior of the mac text system
+            let font = self.font_system.get_font(id).unwrap();
+            let face = self
+                .font_system
+                .db()
+                .faces()
+                .find(|info| info.id == id)
+                .unwrap();
+
+            let font_id = FontId(self.loaded_fonts_store.len());
+            self.loaded_fonts_store.push(font);
+            self.postscript_names
+                .insert(font_id, face.post_script_name.clone());
+
+            font_id
+        }
+    }
+
+    // todo(linux) This is all a quick first pass, maybe we should be using cosmic_text::Buffer
     #[profiling::function]
     fn layout_line(&mut self, text: &str, font_size: Pixels, font_runs: &[FontRun]) -> LineLayout {
         let mut attrs_list = AttrsList::new(Attrs::new());
         let mut offs = 0;
         for run in font_runs {
-            // todo(windows) We need to check we are doing utf properly
-            let font = &self.fonts[run.font_id.0];
+            // todo(linux) We need to check we are doing utf properly
+            let font = &self.loaded_fonts_store[run.font_id.0];
             let font = self.font_system.db().face(font.id()).unwrap();
             attrs_list.add_span(
-                offs..offs + run.len,
+                offs..(offs + run.len),
                 Attrs::new()
                     .family(Family::Name(&font.families.first().unwrap().0))
                     .stretch(font.stretch)
@@ -349,34 +384,32 @@ impl WindowsTextSystemState {
             offs += run.len;
         }
         let mut line = BufferLine::new(text, attrs_list, cosmic_text::Shaping::Advanced);
+
         let layout = line.layout(
             &mut self.font_system,
             font_size.0,
-            f32::MAX, // todo(windows) we don't have a width cause this should technically not be wrapped I believe
+            f32::MAX, // We do our own wrapping
             cosmic_text::Wrap::None,
             None,
         );
         let mut runs = Vec::new();
-        // todo(windows) what I think can happen is layout returns possibly multiple lines which means we should be probably working with it higher up in the text rendering
+
         let layout = layout.first().unwrap();
         for glyph in &layout.glyphs {
             let font_id = glyph.font_id;
-            let font_id = FontId(
-                self.fonts
-                    .iter()
-                    .position(|font| font.id() == font_id)
-                    .unwrap(),
-            );
+            let font_id = self.font_id_for_cosmic_id(font_id);
             let mut glyphs = SmallVec::new();
-            // todo(windows) this is definitely wrong, each glyph in glyphs from cosmic-text is a cluster with one glyph, ShapedRun takes a run of glyphs with the same font and direction
+            // todo(linux) this is definitely wrong, each glyph in glyphs from cosmic-text is a cluster with one glyph, ShapedRun takes a run of glyphs with the same font and direction
             glyphs.push(ShapedGlyph {
                 id: GlyphId(glyph.glyph_id as u32),
                 position: point((glyph.x).into(), glyph.y.into()),
                 index: glyph.start,
                 is_emoji: self.is_emoji(font_id),
             });
+
             runs.push(crate::ShapedRun { font_id, glyphs });
         }
+
         LineLayout {
             font_size,
             width: layout.w.into(),
@@ -446,5 +479,42 @@ impl From<FontStyle> for cosmic_text::Style {
             FontStyle::Italic => cosmic_text::Style::Italic,
             FontStyle::Oblique => cosmic_text::Style::Oblique,
         }
+    }
+}
+
+fn font_into_properties(font: &crate::Font) -> font_kit::properties::Properties {
+    font_kit::properties::Properties {
+        style: match font.style {
+            crate::FontStyle::Normal => font_kit::properties::Style::Normal,
+            crate::FontStyle::Italic => font_kit::properties::Style::Italic,
+            crate::FontStyle::Oblique => font_kit::properties::Style::Oblique,
+        },
+        weight: font_kit::properties::Weight(font.weight.0),
+        stretch: Default::default(),
+    }
+}
+
+fn face_info_into_properties(
+    face_info: &cosmic_text::fontdb::FaceInfo,
+) -> font_kit::properties::Properties {
+    font_kit::properties::Properties {
+        style: match face_info.style {
+            cosmic_text::Style::Normal => font_kit::properties::Style::Normal,
+            cosmic_text::Style::Italic => font_kit::properties::Style::Italic,
+            cosmic_text::Style::Oblique => font_kit::properties::Style::Oblique,
+        },
+        // both libs use the same values for weight
+        weight: font_kit::properties::Weight(face_info.weight.0.into()),
+        stretch: match face_info.stretch {
+            cosmic_text::Stretch::Condensed => font_kit::properties::Stretch::CONDENSED,
+            cosmic_text::Stretch::Expanded => font_kit::properties::Stretch::EXPANDED,
+            cosmic_text::Stretch::ExtraCondensed => font_kit::properties::Stretch::EXTRA_CONDENSED,
+            cosmic_text::Stretch::ExtraExpanded => font_kit::properties::Stretch::EXTRA_EXPANDED,
+            cosmic_text::Stretch::Normal => font_kit::properties::Stretch::NORMAL,
+            cosmic_text::Stretch::SemiCondensed => font_kit::properties::Stretch::SEMI_CONDENSED,
+            cosmic_text::Stretch::SemiExpanded => font_kit::properties::Stretch::SEMI_EXPANDED,
+            cosmic_text::Stretch::UltraCondensed => font_kit::properties::Stretch::ULTRA_CONDENSED,
+            cosmic_text::Stretch::UltraExpanded => font_kit::properties::Stretch::ULTRA_EXPANDED,
+        },
     }
 }
