+++ crates/gpui/src/platform/linux.rs
@@ -1,12 +1,10 @@
 mod client;
 mod dispatcher;
 mod platform;
-mod text_system;
 mod util;
 mod wayland;
 mod x11;
 
 pub(crate) use dispatcher::*;
 pub(crate) use platform::*;
-pub(crate) use text_system::*;
 // pub(crate) use x11::*;
