+++ crates/gpui/src/platform/linux/platform.rs
@@ -24,9 +24,9 @@ use crate::platform::linux::client::Client;
 use crate::platform::linux::wayland::WaylandClient;
 use crate::{
     px, Action, AnyWindowHandle, BackgroundExecutor, ClipboardItem, CursorStyle, DisplayId,
-    ForegroundExecutor, Keymap, LinuxDispatcher, LinuxTextSystem, Menu, PathPromptOptions, Pixels,
-    Platform, PlatformDisplay, PlatformInput, PlatformTextSystem, PlatformWindow, Result,
-    SemanticVersion, Task, WindowOptions, WindowParams,
+    ForegroundExecutor, Keymap, LinuxDispatcher, Menu, PathPromptOptions, Pixels, Platform,
+    PlatformDisplay, PlatformInput, PlatformTextSystem, PlatformWindow, Result, SemanticVersion,
+    Task, WindowOptions, WindowParams,
 };
 
 use super::x11::X11Client;
@@ -57,7 +57,7 @@ pub(crate) struct LinuxPlatformInner {
     pub(crate) loop_signal: LoopSignal,
     pub(crate) background_executor: BackgroundExecutor,
     pub(crate) foreground_executor: ForegroundExecutor,
-    pub(crate) text_system: Arc<LinuxTextSystem>,
+    pub(crate) text_system: Arc<CommonTextSystem>,
     pub(crate) callbacks: RefCell<Callbacks>,
 }
 
@@ -78,7 +78,7 @@ impl LinuxPlatform {
         let use_wayland = wayland_display.is_some_and(|display| !display.is_empty());
 
         let (main_sender, main_receiver) = calloop::channel::channel::<Runnable>();
-        let text_system = Arc::new(LinuxTextSystem::new());
+        let text_system = Arc::new(CommonTextSystem::new());
         let callbacks = RefCell::new(Callbacks::default());
 
         let event_loop = EventLoop::try_new().unwrap();
