+++ crates/gpui/src/platform/windows.rs
@@ -1,14 +1,12 @@
 mod dispatcher;
 mod display;
 mod platform;
-mod text_system;
 mod util;
 mod window;
 
 pub(crate) use dispatcher::*;
 pub(crate) use display::*;
 pub(crate) use platform::*;
-pub(crate) use text_system::*;
 pub(crate) use util::*;
 pub(crate) use window::*;
